server classes moved to communication folder
