+++ server/communication/ServerCommunicator.cpp
