+++ server/communication/ServerCommunicator.h
