+++ server/communication/Stage.cpp
