+++ server/communication/Stage.h
@@ -4,12 +4,12 @@
 #include <vector>
 
 #include "common/Thread.h"
-#include "model/Player.h"
+#include "server/model/Player.h"
 #include "common/StageInfo.h"
 #include "common/Packet.h"
-#include "model/Object.h"
-#include "model/Enemy.h"
-#include "model/Factory.h"
+#include "server/model/Object.h"
+#include "server/model/Enemy.h"
+#include "server/model/Factory.h"
 
 class StageIDProtected {
     private:
